refactoring animations code
