+++ src/artist-page.html
@@ -4,12 +4,14 @@
     <meta charset="UTF-8" />
     <link rel="icon" type="image/svg+xml" href="/favicon.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Artist page</title>
     <meta
       name="description"
       content="The internet's friendliest designer kid."
     />
-    <link rel="stylesheet" href="css/artist-page.css" />
-    <title>Artist page</title>
+    <link rel="stylesheet" href="./css/artist-page.css" />
+    <script type="module" src="./main.js" defer></script>
+    <script type="module" src="./js/artist-page.js" defer></script>
   </head>
   <body>
     <load ="./partials/header.html" />
@@ -603,6 +605,5 @@
     </section>
     <div id="notification" class="notification"></div>
     <load ="./partials/footer.html" />
-    <script type="module" src="./js/artist-page.js"></script>
   </body>
 </html>
