+++ src/css/animations.css
@@ -1,0 +1,63 @@
+/* ANIMATION TRANSFORM Y+*/
+.animated-element.animation-1 {
+  opacity: 0;
+  transform: translateY(20px);
+  transition: opacity 0.9s, transform 0.9s;
+  transition-delay: 300ms;
+}
+
+.animated-element.animation-1.visible {
+  opacity: 1;
+  transform: translateY(0);
+}
+
+/* ANIMATION TRANSFORM X+ */
+.animated-element.animation-2 {
+  opacity: 0;
+  transform: translateX(100px);
+  transition: opacity 0.5s, transform 0.5s;
+  transition-delay: 300ms;
+}
+
+.animated-element.animation-2.visible {
+  opacity: 1;
+  transform: translateX(0);
+}
+
+/* ANIMATION TRANSFORM Y - */
+.animated-element.animation-3 {
+  opacity: 0;
+  transform: translateY(-20px);
+  transition: opacity 0.5s, transform 0.5s;
+  transition-delay: 300ms;
+}
+
+.animated-element.animation-3.visible {
+  opacity: 1;
+  transform: translateY(0);
+}
+
+/* ANIMATION TRANSFORM X- */
+.animated-element.animation-4 {
+  opacity: 0;
+  transform: translateX(-100px);
+  transition: opacity 0.5s, transform 0.5s;
+  transition-delay: 300ms;
+}
+
+.animated-element.animation-4.visible {
+  opacity: 1;
+  transform: translateY(0);
+}
+
+/* ANIMATION SCALE */
+.animated-element.animation-5 {
+  opacity: 0;
+  transform: scale(0);
+  transition: opacity 0.9s, transform 0.9s;
+}
+
+.animated-element.animation-5.visible {
+  opacity: 1;
+  transform: scale(1);
+}
