+++ src/css/styles.css
@@ -7,5 +7,6 @@
 @import url('../../node_modules/modern-normalize/modern-normalize.css');
 @import url('./reset.css');
 @import url('./base.css');
+@import url('./animations.css');
 @import url('./header.css');
 @import url('./footer.css');
