+++ src/js/rankings.js
@@ -1,30 +1,3 @@
-// Animations
-document.addEventListener('DOMContentLoaded', function () {
-    var elements = document.querySelectorAll(
-      '.animated-element'
-    );
-
-  var isInViewport = function (element) {
-    var rect = element.getBoundingClientRect();
-    return (
-      rect.top >= 0 &&
-      rect.bottom <=
-        (window.innerHeight || document.documentElement.clientHeight)
-    );
-  };
-
-  var handleScroll = function () {
-    elements.forEach(function (element) {
-      if (isInViewport(element) && !element.classList.contains('visible')) {
-        element.classList.add('visible');
-      }
-    });
-  };
-
-  document.addEventListener('scroll', handleScroll);
-  handleScroll(); // Викликати функцію на початку для перевірки видимості елементів
-});
-
 // Filter link active
 document.addEventListener('DOMContentLoaded', function () {
   var filterLinks = document.querySelectorAll('.filter-link');
