+++ src/main.js
@@ -1,1 +1,2 @@
-import "./js/menu";
+import './js/animations';
+import './js/menu';
