+++ src/nft-page.html
@@ -10,6 +10,7 @@
     />
     <title>dff-nft-project</title>
     <link rel="stylesheet" href="./css/nft-page.css" />
+    <script type="module" src="./main.js" defer></script>
   </head>
   <body>
     <load ="./partials/header.html" />
@@ -484,6 +485,5 @@
     </main>
 
     <load ="./partials/footer.html" />
-    <script type="module" src="./main.js"></script>
   </body>
 </html>
