+++ src/rankings.html
@@ -10,9 +10,11 @@
     />
     <title>Rankings</title>
     <link rel="stylesheet" href="./css/rankings.css" />
+    <script type="module" src="./main.js" defer></script>
+    <script type="module" src="./js/rankings.js" defer></script>
   </head>
   <body>
-    <load src="partials/header.html" />
+    <load src="./partials/header.html" />
 
     <!-- Write your code here -->
     <section class="section top-creators">
@@ -28,20 +30,56 @@
       <div class="container">
         <h2 class="visually-hidden">Top-20 rankings</h2>
         <div class="filter">
-          <a href="#" class="filter-link filter-link-active animated-element animation-1" data-filter=""
+          <a
+            href="#"
+            class="filter-link filter-link-active animated-element animation-1"
+            data-filter=""
             >Today</a
           >
-          <a href="#" class="filter-link animated-element animation-1" data-filter="">This Week</a>
-          <a href="#" class="filter-link animated-element animation-1" data-filter="">This Month</a>
-          <a href="#" class="filter-link animated-element animation-1" data-filter="">All Time</a>
+          <a
+            href="#"
+            class="filter-link animated-element animation-1"
+            data-filter=""
+            >This Week</a
+          >
+          <a
+            href="#"
+            class="filter-link animated-element animation-1"
+            data-filter=""
+            >This Month</a
+          >
+          <a
+            href="#"
+            class="filter-link animated-element animation-1"
+            data-filter=""
+            >All Time</a
+          >
         </div>
         <div class="filter-mobile">
-          <a href="#" class="filter-link filter-link-active animated-element animation-1" data-filter=""
+          <a
+            href="#"
+            class="filter-link filter-link-active animated-element animation-1"
+            data-filter=""
             >1d</a
           >
-          <a href="#" class="filter-link animated-element animation-1" data-filter="">7d</a>
-          <a href="#" class="filter-link animated-element animation-1" data-filter="">30d</a>
-          <a href="#" class="filter-link animated-element animation-1" data-filter="">All Time</a>
+          <a
+            href="#"
+            class="filter-link animated-element animation-1"
+            data-filter=""
+            >7d</a
+          >
+          <a
+            href="#"
+            class="filter-link animated-element animation-1"
+            data-filter=""
+            >30d</a
+          >
+          <a
+            href="#"
+            class="filter-link animated-element animation-1"
+            data-filter=""
+            >All Time</a
+          >
         </div>
         <div class="table-header animated-element animation-1">
           <span class="number">&#35;</span>
@@ -395,8 +433,6 @@
       </div>
     </section>
 
-    <!-- <load src="partials/footer.html" /> -->
-    <script type="module" src="./js/rankings.js" defer></script>
-    <script src="./js/menu.js" defer></script>
+    <load src="partials/footer.html" />
   </body>
 </html>
